Fix warning about comparing signed and unsigned integers.
The explicit constructor for the DisjointSets class used int in it's
for loop instead of size_t, the type of argument the constructor is expecting.
Trivial fix.

+++ DisjSets.h
@@ -16,7 +16,7 @@
 class DisjointSets {
  public:
   explicit DisjointSets(size_t num_elements) {
-    for (int i = 0; i < num_elements; ++i)
+    for (size_t i = 0; i < num_elements; ++i)
       the_sets_.push_back(-1);
   }
